Add bunch of authorization guard in CRUD controller
feat:
- add authorization in categories, notifications, and users controller

+++ backend/controllers/API/categories/destroy.js
@@ -18,6 +18,14 @@ async function controller(req, res, next){
     })
   }
 
+  if(!req.isAdmin) {
+    return res.json({
+      error: true,
+      message: 'Unauthorized access',
+      data: [],
+    })
+  }
+
   const data = await prisma.category.delete({
     where:{
       id
